+++ packages/taro-rn-runner/src/loaders/utils/processAst.ts
@@ -16,7 +16,7 @@ let componentClassName // get app.js class name
 const providerComponentName = 'Provider'
 const taroComponentsRNProviderName = 'TCRNProvider'
 const setStoreFuncName = 'setStore'
-const routerImportDefaultName = 'TaroRouter'
+// const routerImportDefaultName = 'TaroRouter'
 const DEVICE_RATIO = 'deviceRatio'
 
 const taroApis = [
@@ -54,6 +54,39 @@ const superNode = t.expressionStatement(
   )
 )
 
+function getInitRouterAst (pages, configAst) {
+  const routerPages = pages
+    .map(item => {
+      const pagePath = item.startsWith('/') ? item : `/${item}`
+      const screenName = _.camelCase(pagePath)
+      return `['${item}',${screenName}]`
+    })
+    .join(',')
+
+  return t.expressionStatement(
+    t.assignmentExpression(
+      '=',
+      t.memberExpression(
+        t.thisExpression(),
+        t.identifier('RootStack'),
+        false
+      ),
+      t.callExpression(
+        t.memberExpression(
+          t.identifier('TaroRouter'),
+          t.identifier('initRouter'),
+          false
+        ),
+        [
+          template(`[${routerPages}]`)().expression,
+          template(`${taroImportDefaultName}`)().expression,
+          configAst[0]
+        ]
+      )
+    )
+  )
+}
+
 function getInitPxTransformNode ({designWidth, deviceRatio}) {
   const pxTransformConfig = {designWidth: designWidth || 750}
 
@@ -64,7 +97,7 @@ function getInitPxTransformNode ({designWidth, deviceRatio}) {
   return initPxTransformNode
 }
 
-function getClassPropertyVisitor ({filePath, pages, iconPaths, isEntryFile}) {
+function getClassPropertyVisitor ({filePath, pages, iconPaths, isEntryFile, configAst}) {
   return (astPath) => {
     const node = astPath.node
     const key = node.key
@@ -123,10 +156,21 @@ function getClassPropertyVisitor ({filePath, pages, iconPaths, isEntryFile}) {
               }
             })
           }
+        },
+        MemberExpression (astPath) {
+          const node = astPath.node
+          // if has this.XXX in config
+          if (t.isThisExpression(node.object)) {
+            // replace this
+            astPath.replaceWith(node.property)
+          }
         }
       })
+      configAst.push(value)
+      astPath.remove()
+    } else {
+      astPath.node.static = 'true'
     }
-    astPath.node.static = 'true'
   }
 }
 
@@ -230,19 +274,20 @@ interface IProcessAstArgs {
  * @param alias
  */
 export function processAst ({
-                               ast,
-                               buildAdapter,
-                               type,
-                               designWidth,
-                               deviceRatio,
-                               sourceFilePath: filePath,
-                               sourceDir,
-                               alias
-                             }: IProcessAstArgs) {
+                              ast,
+                              buildAdapter,
+                              type,
+                              designWidth,
+                              deviceRatio,
+                              sourceFilePath: filePath,
+                              sourceDir,
+                              alias
+                            }: IProcessAstArgs) {
   const isEntryFile = type === PARSE_AST_TYPE.ENTRY
   const styleFiles: string[] = []
   const pages: string[] = [] // app.js 里面的config 配置里面的 pages
   const iconPaths: string[] = [] // app.js 里面的config 配置里面的需要引入的 iconPath
+  const configAst = []
   let hasAddReactImportDefaultName = false
   let providorImportName
   let storeName
@@ -369,7 +414,7 @@ export function processAst ({
         source.value = PACKAGES['@tarojs/components-rn']
       }
     },
-    ClassProperty: getClassPropertyVisitor({filePath, pages, iconPaths, isEntryFile}),
+    ClassProperty: getClassPropertyVisitor({filePath, pages, iconPaths, isEntryFile, configAst}),
     ClassMethod: {
       enter (astPath: NodePath<t.ClassMethod>) {
         const node = astPath.node
@@ -484,7 +529,7 @@ export function processAst ({
                   ${funcBody}
                 </${providorImportName}>`
               }
-              node.body = template(`{return (${funcBody});}`, babylonConfig as any)() as any
+              node.body = template(`{const RootStack = this.RootStack;return (${funcBody});}`, babylonConfig as any)() as any
             }
           },
 
@@ -512,7 +557,11 @@ export function processAst ({
                     'constructor',
                     t.identifier('constructor'),
                     [t.identifier('props'), t.identifier('context')],
-                    t.blockStatement([superNode, additionalConstructorNode] as t.Statement[]),
+                    t.blockStatement([
+                      superNode,
+                      additionalConstructorNode,
+                      getInitRouterAst(pages, configAst)
+                    ] as t.Statement[]),
                     false,
                     false
                   )
@@ -591,21 +640,7 @@ export function processAst ({
           })
 
           // TODO Taro.initRouter  生成 RootStack
-          const routerPages = pages
-            .map(item => {
-              const pagePath = item.startsWith('/') ? item : `/${item}`
-              const screenName = _.camelCase(pagePath)
-              return `['${item}',${screenName}]`
-            })
-            .join(',')
-          node.body.push(template(
-            `const RootStack = ${routerImportDefaultName}.initRouter(
-            [${routerPages}],
-            ${taroImportDefaultName},
-            App.config
-            )`,
-            babylonConfig as any
-          )() as any)
+
           // initNativeApi
           const initNativeApi = template(
             `${taroImportDefaultName}.initNativeApi(${taroImportDefaultName})`,